requirements.txt
Flask==2.2.5
SQLAlchemy==1.4.52
Flask_SQLAlchemy==3.0.3
requests==2.31.0
python-dotenv==1.0.0
geopy==2.4.0
pandas==2.1.1
gunicorn==20.1.0
